.github: Fix link to CONTRIBUTING.md in bug report template
The existing relative link won’t work, as it’s eventually resolved
relative to
`github.com/cloud-carbon-footprint/cloud-carbon-footprint/issues/123`
rather than to the `.github/workflows` directory in git, as I suspect
was intended.

The relative link `../blob/trunk/CONTRIBUTING.md` does work when
resolved relative to the issue URI, but is fairly unintelligible by
someone when they’re previewing an issue or reading the Markdown while
filling in the template.

So I think it makes most sense to use the full URI to `CONTRIBUTING.md`,
which will work in all situations.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>